refactor test helpers within ide-completions
